+++ src/__init__.py
@@ -1,10 +1,10 @@
 from .backtest import single_daily_rets, portfolio_performance
 from .data_source import load_kline, sanity_check
 from .walkforward import walk_forward
-from .engines import ma_cross, macd_cross, rsi_reversal, REFERENCE_ENGINES
+from .engines import ma_cross, macd_cross, rsi_reversal, mytt_macd, REFERENCE_ENGINES
 from .benchmark import evaluate_engine
 
 __all__ = ["single_daily_rets", "portfolio_performance",
            "walk_forward", "load_kline", "sanity_check",
-           "ma_cross", "macd_cross", "rsi_reversal", "REFERENCE_ENGINES",
+           "ma_cross", "macd_cross", "rsi_reversal", "mytt_macd", "REFERENCE_ENGINES",
            "evaluate_engine"]